+++ files/classes/cron/submission.py
@@ -8,9 +8,9 @@ from sqlalchemy.sql.sqltypes import Boolean, Integer, String, Text
 from files.classes.cron.tasks import (RepeatableTask, ScheduledTaskType,
                                       TaskRunContext)
 from files.classes.submission import Submission
-from files.classes.visstate import StateMod
+from files.classes.visstate import StateMod, StateReport, VisibilityState
 from files.helpers.config.const import SUBMISSION_TITLE_LENGTH_MAXIMUM
-from files.helpers.content import ModerationState, body_displayed
+from files.helpers.content import body_displayed
 from files.helpers.lazy import lazy
 from files.helpers.sanitize import filter_emojis_only
 
@@ -173,13 +173,12 @@ class ScheduledSubmissionTask(RepeatableTask):
 		return f"/tasks/scheduled_posts/{self.id}/content"
 	
 	@property
-	def moderation_state(self) -> ModerationState:
-		return ModerationState(
-			removed=False,
-			removed_by_name=None,
+	def visibility_state(self) -> VisibilityState:
+		return VisibilityState(
+			state_mod=StateMod.VISIBLE,
+			state_mod_set_by=None,
+			state_report=StateReport.UNREPORTED,
 			deleted=False, # we only want to show deleted UI color if disabled
-			reports_ignored=False,
-			filtered=False,
 			op_shadowbanned=False,
 			op_id=self.author_id_submission,
 			op_name_safe=self.author_name
